Move helpers into same subfolder as Vulnerability Management Reporting integration tests
